Moved server/events.js to shared folder so it can also be used by the client.
