+++ shared/events.js
